Move baseline authorizations in domain
Part of epic #12142 Baseline v1

Objective: check authorizations as soon as possible.
Some authorizations are verified when entering in domain (with domain services).
Indeed, domain services are exposed through REST api, but other interfaces may be
possible. So REST interface is not a good candidate for authorizations checks.

Other authorizations are verified in adapters when check cannot be done earlier.
Ex: with findById, we have to query database to known which project is involved.
It is only once you have the instance, that you know the project.

For maintenance reasons, a single component is responsible for all domain authorization
rules: BaselineAuthorizations.

Note: this change is only about authorizations on baseline resource. Authorizations
on comparions are out of scope for now.

+++ include/Domain/BaselineService.php
@@ -34,10 +34,17 @@ class BaselineService
     /** @var Clock */
     private $clock;
 
-    public function __construct(BaselineRepository $baseline_repository, Clock $clock)
-    {
+    /** @var BaselineAuthorizations */
+    private $authorizations;
+
+    public function __construct(
+        BaselineRepository $baseline_repository,
+        Clock $clock,
+        BaselineAuthorizations $authorizations
+    ) {
         $this->baseline_repository = $baseline_repository;
         $this->clock               = $clock;
+        $this->authorizations      = $authorizations;
     }
 
     /**
@@ -45,6 +52,11 @@ class BaselineService
      */
     public function create(PFUser $current_user, TransientBaseline $baseline): Baseline
     {
+        if (! $this->authorizations->canCreateBaseline($current_user, $baseline)) {
+            throw new NotAuthorizedException(
+                dgettext('tuleap-baseline', "You are not allowed to create this baseline")
+            );
+        }
         return $this->baseline_repository->add(
             $baseline,
             $current_user,
@@ -52,9 +64,6 @@ class BaselineService
         );
     }
 
-    /**
-     * @throws NotAuthorizedException
-     */
     public function findById(PFUser $current_user, int $id): ?Baseline
     {
         return $this->baseline_repository->findById($current_user, $id);
@@ -65,6 +74,11 @@ class BaselineService
      */
     public function delete(PFUser $current_user, Baseline $baseline)
     {
+        if (! $this->authorizations->canDeleteBaseline($current_user, $baseline)) {
+            throw new NotAuthorizedException(
+                dgettext('tuleap-baseline', "You are not allowed to delete this baseline")
+            );
+        }
         return $this->baseline_repository->delete($baseline, $current_user);
     }
 
@@ -82,6 +96,11 @@ class BaselineService
         int $page_size,
         int $baseline_offset
     ): BaselinesPage {
+        if (! $this->authorizations->canReadBaselinesOnProject($current_user, $project)) {
+            throw new NotAuthorizedException(
+                dgettext('tuleap-baseline', "You are not allowed to read baselines of this project")
+            );
+        }
         $baselines = $this->baseline_repository->findByProject($current_user, $project, $page_size, $baseline_offset);
         $count     = $this->baseline_repository->countByProject($project);
         return new BaselinesPage($baselines, $page_size, $baseline_offset, $count);
